# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "https://boards-api.ashysmoke-79f09cd7.uksouth.azurecontainerapps.io",
      "description": "Boards API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/users/{userId}/notifications/{notificationId}": {
      "delete": {
        "tags": ["notification-controller"],
        "operationId": "deleteNotification",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["notification-controller"],
        "operationId": "toggleRead",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/projects/{id}/members": {
      "patch": {
        "tags": ["project-controller"],
        "operationId": "inviteMembers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUsersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/members/{memberId}": {
      "get": {
        "tags": ["member-controller"],
        "operationId": "getMemberById",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["member-controller"],
        "operationId": "removeMember",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["member-controller"],
        "operationId": "updateMemberRole",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{userId}/notifications": {
      "get": {
        "tags": ["notification-controller"],
        "operationId": "getNotificationsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["notification-controller"],
        "operationId": "deleteAllNotifications",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/projects/{projectId}/members": {
      "get": {
        "tags": ["member-controller"],
        "operationId": "getMembersByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}/members/{userId}": {
      "get": {
        "tags": ["member-controller"],
        "operationId": "getCurrentMember",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "IssueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "assignee": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "type": {
            "type": "string",
            "enum": ["EPIC", "TASK", "BUG"]
          },
          "status": {
            "type": "string",
            "enum": ["TO_DO", "IN_PROGRESS", "DONE", "CANCELED"]
          },
          "priority": {
            "type": "string",
            "enum": ["HIGHEST", "HIGH", "MEDIUM", "LOW", "LOWEST"]
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "dueOn": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "updatedOn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "DEVELOPER", "VIEWER"]
          },
          "joinedOn": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "picture": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          }
        }
      },
      "MemberRequest": {
        "required": ["email", "role"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "DEVELOPER", "VIEWER"]
          }
        }
      },
...
      "ProjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "icon": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          }
        }
      },
...
      "InviteUsersRequest": {
        "required": ["members"],
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRequest"
            }
          }
        }
      },
      "MemberRoleUpdateRequest": {
        "required": ["role"],
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "DEVELOPER", "VIEWER"]
          }
        }
      },
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": ["ADDED_TO_PROJECT", "ASSIGNED_TO_ISSUE"]
          },
          "sender": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "receiver": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "issue": {
            "$ref": "#/components/schemas/IssueResponse"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}